refactor: split types and utils
